requirements.txt
mysql-connector-python~=8.3.0
bcrypt~=4.2.0